requirements.txt
Click==7.0
flake8==3.7.5
psycopg2==2.7.7
psycopg2-binary==2.7.7
python-dotenv==0.10.1
Sphinx==1.8.4
